refactor: extract contact search highlighter
