+++ feature/contactsSearch/src/main/java/dev/alenajam/opendialer/feature/contactsSearch/ContactsSearchScreen.kt
@@ -1,7 +1,6 @@
 package dev.alenajam.opendialer.feature.contactsSearch
 
 import android.app.Activity
-import android.content.Context
 import android.provider.ContactsContract
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
@@ -58,13 +57,9 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.TextRange
-import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -75,8 +70,6 @@ import dev.alenajam.opendialer.core.common.telecom.CallAccount
 import dev.alenajam.opendialer.core.common.telecom.CallPlacementResult
 import dev.alenajam.opendialer.core.common.ui.CallAccountPicker
 import dev.alenajam.opendialer.core.common.ui.Dialpad
-import dev.alenajam.opendialer.core.aosp.SmartDialNameMatcher
-import dev.alenajam.opendialer.core.aosp.SmartDialMatchPosition
 import dev.alenajam.opendialer.data.contactsSearch.DialerSearchContact
 
 @Composable
@@ -393,14 +386,14 @@ private fun ResultRow(
         stringResource(ContactsContract.CommonDataKinds.Phone.getTypeLabelResource(contact.phoneType))
     }
     val title = if (contact.name.isNotBlank()) contact.name else contact.number
-    val highlightedTitle = highlightedMatch(
+    val highlightedTitle = highlightSearchMatch(
         context = context,
         text = title,
         query = query,
         isDialpadSearch = isDialpadSearch,
         isPhoneNumber = contact.name.isBlank(),
     )
-    val highlightedNumber = highlightedMatch(
+    val highlightedNumber = highlightSearchMatch(
         context = context,
         text = contact.number,
         query = query,
@@ -504,66 +497,6 @@ private fun ResultRow(
             }
         }
     }
-}
-
-private fun highlightedMatch(
-    context: Context,
-    text: String,
-    query: String,
-    isDialpadSearch: Boolean,
-    isPhoneNumber: Boolean,
-): AnnotatedString {
-    if (text.isBlank() || query.isBlank()) return AnnotatedString(text)
-
-    val matchRanges = if (isDialpadSearch) {
-        val matcher = SmartDialNameMatcher(query)
-        if (isPhoneNumber) {
-            listOfNotNull(matcher.matchesNumber(context, text, query))
-        } else if (matcher.matches(context, text)) {
-            matcher.matchPositions
-        } else {
-            emptyList()
-        }
-    } else if (isPhoneNumber) {
-        textPhoneMatchPositions(text, query)
-    } else {
-        textMatchPositions(text, query)
-    }
-
-    return buildAnnotatedString {
-        append(text)
-        matchRanges.forEach { match ->
-            val start = match.start.coerceIn(0, text.length)
-            val end = match.end.coerceIn(start, text.length)
-            if (start < end) {
-                addStyle(SpanStyle(fontWeight = FontWeight.Bold), start, end)
-            }
-        }
-    }
-}
-
-private fun textMatchPositions(text: String, query: String): List<SmartDialMatchPosition> = buildList {
-    var startIndex = 0
-    while (startIndex < text.length) {
-        val matchStart = text.indexOf(query, startIndex, ignoreCase = true)
-        if (matchStart < 0) break
-        add(SmartDialMatchPosition(matchStart, matchStart + query.length))
-        startIndex = matchStart + query.length
-    }
-}
-
-private fun textPhoneMatchPositions(text: String, query: String): List<SmartDialMatchPosition> {
-    val queryDigits = query.filter(Char::isDigit)
-    if (queryDigits.isEmpty()) return emptyList()
-
-    val digitIndices = text.indices.filter { text[it].isDigit() }
-    val digits = digitIndices.joinToString(separator = "") { text[it].toString() }
-    val matchStart = digits.indexOf(queryDigits)
-    if (matchStart < 0) return emptyList()
-
-    val start = digitIndices[matchStart]
-    val end = digitIndices[matchStart + queryDigits.lastIndex] + 1
-    return listOf(SmartDialMatchPosition(start, end))
 }
 
 @Composable
